fix(claude): address PR review feedback - Move tail-list.ts from commands/ to lib/ (commands as thin wrappers) - Fix countListFlags missing combined flags like -lf (fall back to opts) - Auto-select warning to stderr (prevents breaking --raw JSONL output) - Add allProjects to findSessions in 2+ active branch
