+++ src/claude/lib/tail-list.ts
